FEATURE(openio_service): check service
 ##### SUMMARY

A new variables is available: `openio_service_checks`

It is used to defined a list of checks to do after the service as been
setup. It could be a uri test, a simple port test, a command test or a
shell test. See `defaults/main.yml` for examples.

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION