FEATURE(openio_service): syslog logger
 ##### SUMMARY

gridinit closes stdout and stderr file descriptors for the processes it
forks. For those services that only log to stdout, this PR adds a
python3 script which launch a process by redirecting its stdout to
syslog. this way we can log those processes that are stuck to stdout
logging.

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION