BUGFIX/MINOR(openio-service): set LANG env for syslog services
 ##### SUMMARY

`/usr/bin/gridinit-syslog-logger` use `LANG` environment variable to
determine how it should decode logs read from its child process.

With an empty environment, the default was to decode logs as `ascii` and
it could rise an exception with UTF8 logs.

To fix this, this PR adds the `LANG=en_us.UTF8` environment variable to
each services that use the `redirect_stdout_to_syslog` option.

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION